Change package versioning to use hard-coded value.
This is due to the use of github which produces source tarballs which
can be used for package generation. These tarballs are used outside of
git so git versioning is not possible, hence the hard-coded value.
Because of this, other packaging should probably use the source tarball
rather than a dist tarball.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>