all: Remove most in-file ChangeLogs and versioning
In-file ChangeLog information is no longer required in the modern VCS
world, except where a file is imported as a version, eg bootstrap.sh,
gitignore etc.

In-file versioning should only be provided when it helps to narrow down
bugs or problems. So they are valid as script / program version
information which the user can quote when reporting bugs. They are not
required on AutoTools files for example.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>